refactor: Add element.ts (exports PSVGElement type)
