+++ src/element.ts
@@ -1,0 +1,6 @@
+export interface PSVGElement {
+  tagName: string;
+  children: PSVGElement[];
+  attributes: Record<string, string>;
+  innerHTML: string;
+}
